+++ backend/src/main/java/ru/shutovna/moyserf/model/Vip.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
 import java.time.Instant;
 import java.time.LocalDateTime;
 
@@ -16,7 +15,7 @@ import java.time.LocalDateTime;
 public class Vip {
     @Id
     @GeneratedValue
-    private Long id;
+    private Integer id;
 
     @Enumerated(EnumType.STRING)
     private VipType type;
