+++ backend/src/main/java/ru/shutovna/moyserf/payload/response/UserInfoResponse.java
@@ -1,6 +1,5 @@
 package ru.shutovna.moyserf.payload.response;
 
-import com.nimbusds.openid.connect.sdk.claims.UserInfo;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,11 +9,11 @@ import ru.shutovna.moyserf.model.User;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserInfoResponse {
-    private long id;
+    private Integer id;
     private String email;
     private String name;
     private String imageUrl;
-    private Long invitorId;
+    private Integer invitorId;
 
     public static UserInfoResponse from(User user) {
         UserInfoResponse response = new UserInfoResponse();
