+++ backend/src/main/java/ru/shutovna/moyserf/repository/PasswordResetTokenRepository.java
@@ -9,7 +9,7 @@ import ru.shutovna.moyserf.model.User;
 import java.util.Date;
 import java.util.stream.Stream;
 
-public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, Long> {
+public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, Integer> {
 
     PasswordResetToken findByToken(String token);
