+++ backend/src/main/java/ru/shutovna/moyserf/repository/SiteRepository.java
@@ -7,7 +7,7 @@ import ru.shutovna.moyserf.model.User;
 
 import java.util.List;
 
-public interface SiteRepository extends JpaRepository<Site, Long> {
+public interface SiteRepository extends JpaRepository<Site, Integer> {
     List<Site> findByOwnerOrderByCreatedAtDesc(User owner);
 
     @Query("SELECT s FROM Site s WHERE (SELECT COUNT(o) FROM Order o WHERE o.site.id = s.id and o.closed = false ) > 0" +
