+++ backend/src/main/java/ru/shutovna/moyserf/service/ITransactionService.java
@@ -4,13 +4,14 @@ import ru.shutovna.moyserf.model.Transaction;
 import ru.shutovna.moyserf.model.TransactionType;
 import ru.shutovna.moyserf.model.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface ITransactionService {
     List<Transaction> getTransactions();
 
-    Optional<Transaction> getTransaction(long id);
+    Optional<Transaction> getTransaction(int id);
 
-    Transaction createTransaction(TransactionType type, String description, long sum,  User user);
+    Transaction createTransaction(TransactionType type, String description, long sum, User user);
 }
