+++ backend/src/main/java/ru/shutovna/moyserf/util/Constants.java
@@ -1,6 +1,6 @@
 package ru.shutovna.moyserf.util;
 
 public interface Constants {
-    long SYSTEM_USER_ID = 0;
+    int SYSTEM_USER_ID = 0;
     String AVATARS_UPLOAD_DIR = "uploads/avatars";
 }
